requirements.txt
py621 ~= 0.0.2
pyyaml ~= 6.0
